+++ src/detector.py
@@ -17,8 +17,8 @@ import time
 import ogl_viewer.viewer as gl
 import cv_viewer.tracking_viewer as cv_viewer
 import cv_viewer.labels as lab
-import outputs.retrieve_data as rd
-from algorithm import string_to_label 
+import BRAs_VoiceAndVision.src.history as rd
+from utils import string_to_label 
 
 lock = Lock()
 run_signal = False
