+++ src/main.py
@@ -1,14 +1,14 @@
 from scipy.stats import trim_mean
-import speech_to_text
+#import speech_to_text
 import numpy as np
-import algorithm
-import outputs.retrieve_data as retrieve_data
+import utils
+import BRAs_VoiceAndVision.src.history as history
 import detector
 import argparse
 import torch
 import math
 import faulthandler
-import UART
+import uart
 from enum import Enum
 
 faulthandler.enable()
@@ -23,8 +23,8 @@ def find_angle(coordinated_target_list:np.ndarray, mode: Mode = Mode.FIX_THRESHO
         x = trim_mean(coordinated_target_list[:,0], proportiontocut=0.1)
         z = trim_mean(coordinated_target_list[:,2], proportiontocut=0.1)
     elif mode == Mode.FIX_THRESHOLD:
-        x = retrieve_data.get_distance(coordinated_target_list[:,0])
-        z = retrieve_data.get_distance(coordinated_target_list[:,2])
+        x = history.get_distance(coordinated_target_list[:,0])
+        z = history.get_distance(coordinated_target_list[:,2])
     
     angle_rad = math.atan(x / z)
     angle_deg = math.degrees(angle_rad)
@@ -33,7 +33,7 @@ def find_angle(coordinated_target_list:np.ndarray, mode: Mode = Mode.FIX_THRESHO
 def main():
     text = speech_to_text.transcribe_directly()
     print(text)
-    label = algorithm.string_to_label(text)
+    label = utils.string_to_label(text)
     print(label)
 
     parser = argparse.ArgumentParser()
